+++ packages/cheqd-blockchain-api/src/api/index.js
@@ -347,8 +347,6 @@ export class CheqdAPI extends AbstractApiProvider {
       async () => {
         const res = await this.sdk.signer.signAndBroadcast(sender, txJSON, payment, memo ?? '');
         if (res.code) {
-          console.error(res);
-
           throw new Error(
             JSON.stringify(res, (_, value) => (typeof value === 'bigint' ? `${value.toString()}n` : value)),
           );